1. When extrat publish time, just extract the first one, it seems the last one extracted currently - Testing 2. Always crawl detail pages from seed pages - Testing 3. Add tieba - add them topic by topic - Inject manully 4. Add a filter to filter out old url date

+++ src/java/org/apache/nutch/fetch/FetchMonitor.java
@@ -208,7 +208,9 @@ public class FetchMonitor {
         taskSchedulers.remove(taskScheduler.getId());
       }
 
-      Files.delete(Paths.get(finishScript));
+      if (Files.exists(Paths.get(finishScript))) {
+        Files.delete(Paths.get(finishScript));
+      }
     }
     catch (Throwable e) {
       LOG.error(StringUtil.stringifyException(e));
